Merge dev to master

See merge request Dkorolkov/mmtback!11

+++ explorer-personal-cabinet-service/src/main/java/org/example/service/CourseProgressService.java
@@ -10,14 +10,16 @@ import org.example.dto.starsystem.StarSystemWithDependenciesGetResponse;
 import org.example.dto.starsystem.SystemDependencyModel;
 import org.example.exception.classes.courseEX.CourseNotFoundException;
 import org.example.exception.classes.explorerEX.ExplorerNotFoundException;
+import org.example.exception.classes.progressEX.CourseAlreadyCompletedException;
 import org.example.model.Explorer;
 import org.example.model.Person;
 import org.example.model.course.Course;
 import org.example.model.course.CourseTheme;
 import org.example.model.courserequest.CourseRegistrationRequest;
-import org.example.repository.*;
+import org.example.repository.ExplorerRepository;
 import org.example.repository.course.CourseRepository;
 import org.example.repository.course.CourseThemeRepository;
+import org.example.repository.courseprogress.CourseMarkRepository;
 import org.example.repository.courseprogress.CourseThemeCompletionRepository;
 import org.example.repository.courseregistration.CourseRegistrationRequestRepository;
 import org.example.repository.custom.StarSystemRepository;
@@ -37,6 +39,7 @@ public class CourseProgressService {
     private final CourseThemeRepository courseThemeRepository;
     private final CourseRegistrationRequestRepository courseRegistrationRequestRepository;
     private final StarSystemRepository starSystemRepository;
+    private final CourseMarkRepository courseMarkRepository;
 
     private final KafkaTemplate<String, Integer> kafkaTemplate;
 
@@ -127,6 +130,8 @@ public class CourseProgressService {
         Optional<Explorer> explorer = explorerRepository.findExplorerByPersonIdAndCourseId(personId, courseId);
         if (explorer.isEmpty())
             throw new ExplorerNotFoundException(courseId);
+        if (courseMarkRepository.existsById(explorer.get().getExplorerId()))
+            throw new CourseAlreadyCompletedException(courseId);
         explorerRepository.deleteById(explorer.get().getExplorerId());
         CourseRegistrationRequest request = courseRegistrationRequestRepository
                 .findApprovedCourseRegistrationRequestByPersonIdAndCourseId(personId, courseId);
